% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{Check Installed Packages}
\usage{
check_installed_packages(pkg_status)
}
\arguments{
\item{pkg_status}{Data frame tracking package status.}
}
\value{
Updated pkg_status data frame.
}
\description{
Helper function to check which packages are already installed.
}
\keyword{internal}
