% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{generate_markdown_template_function}
\alias{generate_markdown_template_function}
\title{Generate Template Function for Quarto or R Markdown Files}
\usage{
generate_markdown_template_function(
  file_content,
  template_name,
  description,
  file_ext
)
}
\arguments{
\item{file_content}{Character vector containing the content of the source file.}

\item{template_name}{Character string. The name for the new template function.}

\item{description}{Character string. A brief description of the template.}

\item{file_ext}{Character string. The file extension of the source file ("qmd" or "rmd").}
}
\value{
Character string containing the R code for the new template creation function.
}
\description{
Helper function to generate template creation code for Quarto or R Markdown files.
}
\keyword{internal}
