% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRproject-project.R
\docType{package}
\name{initRproject-package}
\alias{initRproject}
\alias{initRproject-package}
\title{initRproject: Automatically loaded functions when an R project is launched}
\description{
This packages includes a series of functions that could be loaded When an R project is launched.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenyu-psy/initRproject}
  \item Report bugs at \url{https://github.com/chenyu-psy/initRproject/issues}
}

}
\author{
\strong{Maintainer}: Chenyu Li \email{chenyu.li.hz@gmail.com}

}
\keyword{internal}
