% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{install_cran_packages}
\alias{install_cran_packages}
\title{Install CRAN Packages}
\usage{
install_cran_packages(pkg_status, cran_pkgs, auto_install)
}
\arguments{
\item{pkg_status}{Data frame tracking package status.}

\item{cran_pkgs}{Character vector of CRAN package identifiers.}

\item{auto_install}{Logical. Whether to install without prompting.}
}
\value{
Updated pkg_status data frame.
}
\description{
Helper function to install packages from CRAN.
}
\keyword{internal}
