% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{install_github_packages}
\alias{install_github_packages}
\title{Install GitHub Packages}
\usage{
install_github_packages(pkg_status, github_pkgs, auto_install, github_auth)
}
\arguments{
\item{pkg_status}{Data frame tracking package status.}

\item{github_pkgs}{Character vector of GitHub package identifiers.}

\item{auto_install}{Logical. Whether to install without prompting.}

\item{github_auth}{Character. GitHub authentication token.}
}
\value{
Updated pkg_status data frame.
}
\description{
Helper function to install packages from GitHub.
}
\keyword{internal}
