% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{Install Missing Packages}
\usage{
install_missing_packages(pkg_status, auto_install, github_auth)
}
\arguments{
\item{pkg_status}{Data frame tracking package status.}

\item{auto_install}{Logical. Whether to install without prompting.}

\item{github_auth}{Character. GitHub authentication token.}
}
\value{
Updated pkg_status data frame.
}
\description{
Helper function to install packages that are not yet installed.
}
\keyword{internal}
