% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{interactive_package_manager}
\alias{interactive_package_manager}
\title{Interactively or Automatically Install/Update Packages}
\usage{
interactive_package_manager(
  check_df,
  cran_repos = "https://cran.rstudio.com/",
  quietly = FALSE
)
}
\arguments{
\item{check_df}{Data frame as returned by check_installed_packages().}

\item{cran_repos}{CRAN repository URL.}

\item{quietly}{Logical, if TRUE, skip all prompts and install/update automatically.}
}
\value{
Data frame of actions taken.
}
\description{
Interactively or Automatically Install/Update Packages
}
