% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load Packages}
\usage{
load_packages(pkg_status)
}
\arguments{
\item{pkg_status}{Data frame tracking package status.}
}
\value{
Updated pkg_status data frame.
}
\description{
Helper function to load installed packages.
}
\keyword{internal}
