% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{load_parsed_packages}
\alias{load_parsed_packages}
\title{Load All Packages from Parsed Data Frame}
\usage{
load_parsed_packages(parsed_df, quietly = FALSE)
}
\arguments{
\item{parsed_df}{Data frame as returned by parse_packages().}

\item{quietly}{Logical, suppress messages if TRUE.}
}
\value{
Invisibly returns a vector of loaded package names.
}
\description{
Load All Packages from Parsed Data Frame
}
