% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_symlink.R
\name{move_directory_contents}
\alias{move_directory_contents}
\title{Move Contents from One Directory to Another}
\usage{
move_directory_contents(source_dir, target_dir, merge = FALSE)
}
\arguments{
\item{source_dir}{Character string. The source directory.}

\item{target_dir}{Character string. The target directory.}

\item{merge}{Logical. If TRUE, files with the same name will be skipped with a warning.
If FALSE, the function will stop if there are conflicts.}
}
\value{
Invisibly returns TRUE if successful, FALSE otherwise.
}
\description{
Move Contents from One Directory to Another
}
\keyword{internal}
