% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_symlink.R
\name{create_symlink}
\alias{create_symlink}
\title{Create a Symbolic Link to a Directory}
\usage{
create_symlink(target_path, link_path)
}
\arguments{
\item{target_path}{Character string. The target directory path that the symlink will point to.}

\item{link_path}{Character string. The path where the symbolic link will be created.}
}
\value{
Invisibly returns TRUE if the symlink was created successfully, FALSE otherwise.
}
\description{
Creates a symbolic link from a link path to a target directory path.
On Windows, this function will provide instructions for creating symbolic links
as they require administrator privileges and different commands.
}
\details{
On Unix-like systems (macOS, Linux), this function creates a symbolic link using the 'ln -s' command.
On Windows, symbolic links require administrator privileges and use different commands, so this
function will provide instructions for manual creation.
}
\examples{
\dontrun{
create_symlink("~/data_storage", "~/data")
}

}
